requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
jinja2>=3.1.0
google-generativeai>=0.3.0
anthropic>=0.7.0
pytest>=7.4.0
